freelance-crawler/requirements.txt
# Web scraping dependencies
selenium==4.15.2
beautifulsoup4==4.12.2
undetected-chromedriver==3.5.4

# Database
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
requests==2.31.0
